pyproject.toml
[tool.poetry]
name = "djangoapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.9"
django = "^5.1.2"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.10"
drf-yasg = "^1.21.8"
djangorestframework = "^3.15.2"
cloudinary = "^1.41.0"
django-cloudinary-storage = "^0.3.0"
django-compressor = "^4.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
